# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/v1/users/{id}/avatar:
    get:
      tags:
      - user-rest-controller
      operationId: getUserAvatar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
      - user-rest-controller
      operationId: updateUserAvatar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "200":
          description: OK
  /api/v1/teams/{id}:
    get:
      tags:
      - team-rest-controller
      operationId: getTeam
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamDTO"
    put:
      tags:
      - team-rest-controller
      operationId: updateTeam
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Team"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamDTO"
    delete:
      tags:
      - team-rest-controller
      operationId: deleteTeam
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/ratings/{id}:
    get:
      tags:
      - rating-rest-controller
      operationId: getRatingById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RatingDTO"
    put:
      tags:
      - rating-rest-controller
      operationId: updateRating
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RatingDTO"
    delete:
      tags:
      - rating-rest-controller
      operationId: deleteRating
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/matches/{id}:
    get:
      tags:
      - match-rest-controller
      operationId: getMatch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchDTO"
    put:
      tags:
      - match-rest-controller
      operationId: updateMatch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Match"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchDTO"
    delete:
      tags:
      - match-rest-controller
      operationId: deleteMatch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/events/{id}:
    get:
      tags:
      - match-event-rest-controller
      operationId: getMatchEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchEventDTO"
    put:
      tags:
      - match-event-rest-controller
      operationId: updateMatchEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchEventDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchEventDTO"
    delete:
      tags:
      - match-event-rest-controller
      operationId: deleteMatchEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/users/register:
    post:
      tags:
      - user-rest-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistrationDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
  /api/v1/users/login:
    post:
      tags:
      - user-rest-controller
      operationId: login
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
  /api/v1/teams:
    get:
      tags:
      - team-rest-controller
      operationId: getTeams
      parameters:
      - name: page
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageTeamDTO"
    post:
      tags:
      - team-rest-controller
      operationId: createTeam
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Team"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamDTO"
  /api/v1/matches:
    get:
      tags:
      - match-rest-controller
      operationId: getMatches
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageMatchDTO"
    post:
      tags:
      - match-rest-controller
      operationId: createMatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Match"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchDTO"
  /api/v1/matches/{matchId}/events:
    get:
      tags:
      - match-event-rest-controller
      operationId: getEventsByMatch
      parameters:
      - name: matchId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchEventDTO"
    post:
      tags:
      - match-event-rest-controller
      operationId: createMatchEvent
      parameters:
      - name: matchId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchEventDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchEventDTO"
  /api/v1/events/{eventId}/ratings:
    post:
      tags:
      - rating-rest-controller
      operationId: createRating
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RatingDTO"
  /admin/matches/match/save:
    post:
      tags:
      - match-controller
      operationId: saveMatch
      parameters:
      - name: match
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Match"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/users:
    get:
      tags:
      - user-rest-controller
      operationId: getAllUsers
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDTO"
  /api/v1/users/{id}:
    get:
      tags:
      - user-rest-controller
      operationId: getUserByUsername
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
    delete:
      tags:
      - user-rest-controller
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/users/me:
    get:
      tags:
      - user-rest-controller
      operationId: me
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
  /api/v1/teams/{id}/logo:
    get:
      tags:
      - team-rest-controller
      operationId: getTeamLogo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
  /api/v1/ratings:
    get:
      tags:
      - rating-rest-controller
      operationId: getAllRatings
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageRatingDTO"
  /api/v1/events:
    get:
      tags:
      - match-event-rest-controller
      operationId: getAllMatchEvents
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageMatchEventDTO"
components:
  schemas:
    Team:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        stadiumName:
          type: string
        points:
          type: integer
          format: int32
        playedMatchs:
          type: integer
          format: int32
        wins:
          type: integer
          format: int32
        draws:
          type: integer
          format: int32
        losses:
          type: integer
          format: int32
        logoData:
          type: object
          properties:
            binaryStream: {}
        playedMatches:
          type: integer
          format: int32
    TeamDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        stadiumName:
          type: string
        points:
          type: integer
          format: int32
    RatingDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        score:
          type: integer
          format: int32
          maximum: 10
          minimum: 1
        comment:
          type: string
          minLength: 1
        createdAt:
          type: string
          format: date-time
        authorUsername:
          type: string
        authorId:
          type: integer
          format: int64
        eventId:
          type: integer
          format: int64
      required:
      - comment
      - eventId
      - score
    Match:
      type: object
      properties:
        id:
          type: integer
          format: int64
        localTeam:
          $ref: "#/components/schemas/Team"
        visitorTeam:
          $ref: "#/components/schemas/Team"
        localGoals:
          type: integer
          format: int32
        visitorGoals:
          type: integer
          format: int32
        weather:
          type: string
        stadium:
          type: string
        matchDate:
          type: string
          format: date
        matchTime:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/MatchEvent"
        played:
          type: boolean
        climaLluvia:
          type: boolean
        climaDespejado:
          type: boolean
        climaViento:
          type: boolean
        climaNublado:
          type: boolean
        finalized:
          type: boolean
        climaNieve:
          type: boolean
    MatchEvent:
      type: object
      properties:
        id:
          type: integer
          format: int64
        minute:
          type: integer
          format: int32
        type:
          type: string
        namePlayer:
          type: string
        namePlayerOut:
          type: string
        namePlayerIn:
          type: string
        match:
          $ref: "#/components/schemas/Match"
        team:
          $ref: "#/components/schemas/Team"
        ratings:
          type: array
          items:
            $ref: "#/components/schemas/Rating"
        icon:
          type: string
        description:
          type: string
        isGoal:
          type: boolean
        isLocal:
          type: boolean
        isRed:
          type: boolean
        isYellow:
          type: boolean
        isVisitor:
          type: boolean
        isSubstitution:
          type: boolean
    Rating:
      type: object
      properties:
        id:
          type: integer
          format: int64
        score:
          type: integer
          format: int32
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/User"
        event:
          $ref: "#/components/schemas/MatchEvent"
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        avatarData:
          type: object
          properties:
            binaryStream: {}
        ratings:
          type: array
          items:
            $ref: "#/components/schemas/Rating"
        favouriteTeam:
          $ref: "#/components/schemas/Team"
        password:
          type: string
    MatchDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        localTeamName:
          type: string
        visitorTeamName:
          type: string
        localGoals:
          type: integer
          format: int32
        visitorGoals:
          type: integer
          format: int32
        stadium:
          type: string
        date:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/MatchEventDTO"
    MatchEventDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        minute:
          type: integer
          format: int32
          maximum: 120
          minimum: 0
        type:
          type: string
          minLength: 1
        namePlayer:
          type: string
        namePlayerOut:
          type: string
        namePlayerIn:
          type: string
        matchId:
          type: integer
          format: int64
        teamId:
          type: integer
          format: int64
        teamName:
          type: string
      required:
      - matchId
      - minute
      - teamId
      - type
    UserRegistrationDTO:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        favouriteTeamId:
          type: integer
          format: int64
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        favouriteTeamId:
          type: integer
          format: int64
        roles:
          type: array
          items:
            type: string
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageTeamDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/TeamDTO"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageRatingDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/RatingDTO"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageMatchDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/MatchDTO"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageMatchEventDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/MatchEventDTO"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean